python-server/requirements.txt
flask==2.0.1
faiss-cpu==1.9.0
python-dotenv==0.19.2
numpy==1.24.3
chromadb
langchain==0.1.4
tiktoken==0.7.0
openai==1.3.4
requests==2.31.0